extract files to require.
